requirements.txt
telegram>=0.0.1
Telethon>=1.24.0
pandas>=1.4.0
mysql-connector-python>=8.0.28